UC_AccesClient : regler les problemes dans la   methode update

+++ src/main/java/tgpr/bank/model/Account.java
@@ -1,5 +1,6 @@
 package tgpr.bank.model;
 
+import tgpr.framework.Error;
 import tgpr.framework.Model;
 import tgpr.framework.Params;
 
@@ -12,7 +13,8 @@ import java.util.List;
 import java.sql.*;
 import java.util.Objects;
 
-public class Account extends Model{
+public class
+Account extends Model{
 
     private int id;
     private String iban;
@@ -168,20 +170,20 @@ public class Account extends Model{
         return c == 1;
     }
 
-    public void deleteAccess(int id ,int accountId) {
-        if(numberofhorder(accountId)==1) {
-            execute("delete from access where access.type=\"proxy\" and  account=:accountId and user=:userId", new Params("userId", id).add("accountId", accountId));
+    public void deleteAccess(int id ,int accountId,String type) {
+        if(numberofhorder(accountId)==1 && type== "holder") {
+            execute("delete from access where access.type=:type and  account=:accountId and user=:userId", new Params("userId", id).add("accountId", accountId).add("type",type));
         }
         else{
-            execute("delete from access where account=:accountId and user=:userId", new Params("userId", id).add("accountId", accountId));
+            execute("delete from access where  access.type=:type and account=:accountId and user=:userId", new Params("userId", id).add("accountId", accountId));
 
         }
     }
     public void update(int accountId, String type){
         if(numberofhorder(accountId)>1){
             execute("update access set type=:type where access.account=:accountId",new Params(":type",type).add("accountId",accountId) );
-        }
-    }
+
+    }}
     public Integer numberofhorder(int accountid){
        return queryScalar(Integer.class,"select count (access.user),access.account from access where access.type ='holder' and access.account=:accountid " +
                "group by access.account " +
@@ -234,21 +236,21 @@ public class Account extends Model{
 
     }
     public void updateAccess(int accountId,int userId,String type){
-        if(numberOfHolder(accountId)>1){
-            update(accountId,userId,type);
+        if(numberOfHolder(accountId)>1) {
+            update(accountId, userId, type);
         }
 
     }
 
     public void update(int accountId,int userId,String type) {
-        String sql = "update access set type=:type where access.account=:accountId and userId:=userId";
+        String sql = "update access set type=:type where access.account=:accountId and access.user=:userId";
         execute(sql, new Params()
                 .add("accountId",accountId)
                 .add("userId", userId)
                 .add("type",type));
     }
 
-    public  Integer numberOfHolder(int accountId){
+    public  int numberOfHolder(int accountId){
        return queryScalar(Integer.class,"select count(*) from access where access.type='holder' and access.account=:accountId group by access.account",new Params("accountId",accountId));
 
     }
